changed va datatype for getting MIN

+++ srcs/get_values.c
@@ -6,7 +6,7 @@
 /*   By: thuynh <thuynh@student.42.us.org>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/04 21:05:12 by thuynh            #+#    #+#             */
-/*   Updated: 2017/08/14 12:03:10 by thuynh           ###   ########.fr       */
+/*   Updated: 2017/08/14 12:06:04 by thuynh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 void	get_values(t_print **list, va_list ap)
 {
 	if (f_exists(list, '*') && !P)
-		MIN = va_arg(ap, int);
+		MIN = va_arg(ap, ssize_t);
 	if (f_exists(list, '*') && !PNB && P)
-		PNB = va_arg(ap, size_t);
+		PNB = va_arg(ap, ssize_t);
 	if (F == 'c' || F == 'd' || F == 'i' || F == '%' || F == 'D')
 		VAL_INT = va_arg(ap, intmax_t);
 	else if (F == 'u' || F == 'U' || F == 'o' || F == 'O' || F == 'x' ||
